backend/requirements.txt
Django==5.1.5
django-jazzmin==3.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.3
python-dotenv==1.0.1
django-cors-headers==4.6.0
gunicorn==23.0.0
reportlab==4.2.5
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
